Made ship lighter
Ok now it updates based on ship_power and ship_falling_speed so thats
nice. Also it falls and rises at around the same speed but it kinda
feels like the real one as in like it takes longer to start going down
then going up if that makes sense? I don't really know how to explain it
so just look at the code or play it yourself idk.