Replaced the point removal system in player path
I'm stupid so I made it so points got deleted when off screen. Sounds
good right? No! That made the path almost useless and made it so you
couldn't see like any of it while in the editor. So I replaced it with
just not rendering the off screen ones.