Reset player cam y when entering the level
Why didn't it do this before? Well thats a classic question! I
apparently was very slow in the head a while ago! Considering I used -
instead of + in the top hitbox for the block which caused multiple bugs
and stressful nights just for me to randomly change one character in the
code and it works fine!