Added password to the upload level request
This just makes it more secure. Also yes I know it automatically
authenticates you when you start the game AND it checks if your logged
in before uploading the level but people can still edit the source. I
didn't feel like making sure that the level didn't upload if the pass is
wrong but I'll test it now ig.