Fixed no touch and hide being broken
So apparently ever since I added no touch and hide back in 1.3 if no
touch was on an object it just wouldn't save and if hide was on an
object it would save but no touch would also be on that object. This was
easy to fix but it took me until I tried adding groups to notice this
somehow. Sorry about this. Fix will be in 1.6